requirements.txt.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
yfinance>=0.1.70
joblib>=1.1.0
pyspark>=3.2.0
mplfinance>=0.12.7b0
streamlit>=1.10.0
plotly>=5.5.0
requests>=2.25.0
websocket-client>=1.0.0
